requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
charset-normalizer==3.3.2
Django==4.2.14
django-allauth==0.63.6
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-mssql-backend==2.8.1
django-select2==8.2.0
django-widget-tweaks==1.5.0
Jinja2==3.1.4
mssql-django==1.5
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.7
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2

